Create new db copy each time tests are run
