+++ server/test/util/resetDb.ts
@@ -12,14 +12,6 @@ const pgport = process.env.POSTGRES_PORT
   ? Number(process.env.POSTGRES_PORT)
   : 5432;
 
-const dbConfig = {
-  host: pghost,
-  port: pgport,
-  user: pguser,
-  password: pgpass,
-  database: pgdb,
-};
-
 const dbConfigClean = {
   host: pghost,
   port: pgport,
@@ -28,24 +20,7 @@ const dbConfigClean = {
   database: 'postgres_copy',
 };
 
-/** Creates a copy of the current database. Must be run before getCookies() */
-export const setupDb = async (): Promise<void> => {
-  const client = new Client(dbConfig);
-  await client.connect();
-
-  const dbQuery = await client.query(
-    "SELECT FROM pg_database WHERE datname = 'postgres_copy'"
-  );
-  if (dbQuery.rowCount === 0) {
-    await client.query(
-      `CREATE DATABASE postgres_copy WITH TEMPLATE postgres OWNER ${pguser}`
-    );
-  }
-
-  await client.end();
-};
-
-export const cleanDb = async (): Promise<void> => {
+export const resetDb = async (): Promise<void> => {
   // Remove current database and replace with copy
   const client = new Client(dbConfigClean);
   await client.connect();
